add following_list helper

+++ app/helpers/application_helper.rb
@@ -24,18 +24,38 @@ module ApplicationHelper
   end
 
   def print_follow_btn(user)
-    if following?(user)
-      content_tag :div, class: 'follow-btn f-profile' do
-        link_to(following_path(id: user.id), method: :delete) do
-          image_tag 'https://image.flaticon.com/icons/svg/149/149146.svg', size: '40x40'
+    unless current_user?(user)
+      if following?(user)
+        content_tag :div, class: 'follow-btn f-profile' do
+          link_to(following_path(id: user.id), method: :delete) do
+            image_tag 'https://image.flaticon.com/icons/svg/149/149146.svg', size: '40x40'
+          end
         end
-      end
-    else
-      content_tag :div, class: 'follow-btn f-profile' do
-        link_to(followings_path(id: user.id), method: :post) do
-          image_tag 'https://image.flaticon.com/icons/svg/149/149145.svg', size: '40x40'
+      else
+        content_tag :div, class: 'follow-btn f-profile' do
+          link_to(followings_path(id: user.id), method: :post) do
+            image_tag 'https://image.flaticon.com/icons/svg/149/149145.svg', size: '40x40'
+          end
         end
       end
     end
+  end
+
+  def print_cover(user)
+    if user.cover.attachment 
+      image_tag(user.cover, class: 'w-100') 
+    else
+      image_tag('cover.jpg', class: 'w-100')
+    end
+  end
+
+  def render_following_users(user_following_list)
+
+    if !user_following_list.empty?
+        render 'users/follows', list:  user_following_list
+    else
+        "this user has no followers yet."
+    end
+
   end
 end
